Fix dport and port iteration
The current implementation just checks that the port provided by the user is iterable.
This results in a value such as:

dport: 1234-5678

Being iterated over and trying to apply "1,2,3,4,-,5,6,7,8" which is obviously not desireable.
This change adds more strict checking to the value to ensure we only try to iterate over lists.

Resolves: https://issues.redhat.com/browse/OSPRH-7633
Signed-off-by: Brendan Shephard <bshephar@redhat.com>