Disable martian packts logging as default
Logging martian packets is done via printk. Depending on the default
console configuration, it may cause writing on a serial line. Some UART
controllers are slow, and it may stall CPUs for up to 200ms. When this
logging happens, it causes random latency issues.

Disable logging of martian packets by default. It can be enabled for
debugging purposes.