nftables: disable logging of dropped packets by default
Depending on the console redirection configured on boot, logging packets
before dropping them may cause excessive latency issues. Sometimes this
can even stall a CPU for 200ms with some UART controllers and
console=ttyS0.

In most openstack deployments, the is a bridge that may receive random
DHCP requests. For example, when creating a lot of SRIOV VFs,
NetworkManager will try to configure them with DHCP and broadcast
packets. This causes the packets to be received by all overcloud nodes
and be dropped, e.g:

 kernel: IN=ens8f0v2 OUT= MAC=ff:ff:ff:ff:ff:ff:2a:75:ff:3d:02:f0:08:00
 SRC=0.0.0.0 DST=255.255.255.255 LEN=327 TOS=0x00 PREC=0xC0 TTL=64 ID=0
 DF PROTO=UDP SPT=68 DPT=67 LEN=307

I could not find any motivation for the existence of this LOG rule. Not
event in puppet tripleo where this log rule was first introduced in
2015.

I think this rule has no real purpose and should be removed. There are
counters for dropped packets in nftables. And if users want to know what
packets are dropped, they can see them using tcpdump.

The LOG rule can be restored using the edpm_nftables_debug role
variable.

Link: https://github.com/stackforge/puppet-openstack-cloud/commit/5102e5130ae4
Signed-off-by: Robin Jarry <rjarry@redhat.com>