tripleo_cleanup: Fix a task that was always skipped and it shouldn't be
Apparently the 'when' condition on a block is not evaluated only
before the block begins executing, but it is evaluated for each task
separately. That means that if the fact that is used in the 'when'
condition changes during the block execution, it affects whether
the following tasks in the block will execute or not.

In our case, the "Filter for tripleo services" task set the
tripleo_services fact, and due to that the following "Filter services
in skip list" task never executed because the "when" condition on the
block says to execute only if tripleo_services is empty.

I've pulled the two set_fact tasks into a single one to get around
this issue, since i didn't want to pollute the facts list with a
temporary fact. (Facts can't be undefined.)

Co-Authored-By: Joan F. Gilabert Navarro <jgilaber@redhat.com>