[edpm_kernel] Improved kernel args block conditions
The conditions required to enter the kernel args configuration block
have been refactored.

The regular expression that checks for the `edpm_kernel_args` variable value
presence in the command line arguments has been rewritten to intercept the substrings of the same parameters.

Also, the molecule test for this particular scenario, `kernelargs-update-substring` has been rewritten to catch these kind of errors.