Remove role dependency on gathered_facts
The FRR Metadata currently depends on gathered_facts as per the
referenced Jira. Accordingly, the role fails to deploy since we have
remove the need to always gather_facts. This change updates the default
value in meta argument_specs so that we can default to "" when no facts
have been gsthered. This allows us to instead pick up the
variable during the execution of the role using the `setup` module.

Jira: https://issues.redhat.com/browse/OSPRH-10338
Signed-off-by: Brendan Shephard <bshephar@redhat.com>