Fix CAcert mounting
The logic to determine whether to mount a cacert bundle was
mistakenly placed in configure.yml instead of run.yaml where the
templates that use this logic are generated. (run.yaml and
run_ovn.yaml)  This breaks updates which execute run.yaml instead
of configure.yaml

This patch moves the logic to the right place.
Resolves: https://issues.redhat.com/browse/OSPRH-10450