Fix /var/lib/nova access for NFS shares
When /var/lib/nova is mounted via NFS, Kolla driven permissions
change over that directory breaks access to it for VM instances.

That had been already solved in TripleO by introducing a custom
init container that calls a script for smart ownership changes.

Reintroduce that script and tests from TripleO as is, and add a
Nova EDPM init container to call it instead of recursive chown.

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>