Enable kpatch support for updates
This change enables kpatch-patch support, allowing in-place kernel
patching for security and bug fixes, provided
`kpatch-patch-KERNEL_VERSION` package exists.

For more information about kpatch:
https://docs.redhat.com/en/documentation/red_hat_enterprise_linux/9/html/managing_monitoring_and_updating_the_kernel/applying-patches-with-kernel-live-patching_managing-monitoring-and-updating-the-kernel

This PR also corrects a potential issue related to the exclide list
applied to the package update:

if a user were to pass the `edpm_update_exclude_packages` parameter, the
subsequent override wouldn't be applied, meaning the `openvswitch`
package wouldn't be excluded in the end.
Using an "internal var" such as the `_exclude_packages` in order to
inject our own content in addition to whatever the user may pass ensures
everything is working as expected.

Fixes: https://issues.redhat.com/browse/OSPRH-11274