Use PAM for sshd Motd
This change switches the method we use to display the motd.
Previously, we used PrintMotd in the sshd config. RHEL9 configures
pam_motd.so in PAM, and as such the motd will be duplicated when
using PrintMotd. To avoid this, we can configure the motd via PAM.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=2329414
Signed-off-by: Brendan Shephard <bshephar@redhat.com>