Partial Revert "Use PAM for sshd Motd"
This partially reverts commit e032cf239ff7cf74912ec6308527846b69de4621.
RHEL9 and greater will use PAM by default. We only need to ensure
PrintMotd is set to "no" to avoid duplication of the motd message.

We changed the default for PrintMotd in:
https://github.com/openstack-k8s-operators/edpm-ansible/pull/838

This change ensures that we don't override the PrintMotd value and
instead we allow users to be explicit about this, otherwise, we default
to no.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>