[tripleo_cleanup] Add daemon-reload and reset-failed as last step
In the role which cleans up old tripleo related systemd services this
patch adds call for "systemctl daemon-reload" and "systemctl
reset-failed <service_name>" for each tripleo service which was removed.
This is done to make sure that those services aren't visible in the
systemctl output with e.g. failure that unit file don't exists.

Closes: #OSPRH-11323