Apply configuration tasks for Nova on updates
When 'updates' service runs, it only calls install.yml from Nova
service role, which is not sufficient to provide required configuration
updates. Inlcude configure.yml as well to cover that for Nova service.

To properly address this for *all services* we should add
migration/update tasks there, which can be called individually
from the update playbook.

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>