Re-enable swap for bootc
The default path for a swapfile needs to be under a writeable location
for bootc. Use /var/swap. The value of the existing variable,
edpm_bootstrap_swap_path can't just be changed since there are already
existing deployments, so a new variable, edpm_bootstrap_swap_path_bootc
is added for bootc. An ansible task will select which variable to use
based on the type of system.

Signed-off-by: James Slagle <jslagle@redhat.com>