Always load custom nftables rules
This change ensures the custom nftables rules are always loaded by
removing the conditional check and associated sential file. This is necessary
since we're always loading the iptables compatibility rules and this sets the
policy of the INPUT chain to accept. After our custom rules are configured,
this policy should be reset to drop to ensure we're only allowing defined
traffic and dropping anything that doesn't match a rule in the chain.

Jira: OSPRH-14250
Signed-off-by: Brendan Shephard <bshephar@redhat.com>