Apply updates for edpm_nova role instead of config
Inlcuding configure tasks for nova role during updates
doesn't work as there are no required data sources attached
to osdpd calling the update service.

Introduce the update tasks which contents would be changing
from release to release and include only tasks required
to "migrate" the configuration and initialization logic
of the nova EDPM service from A to B, then from B to C
and so on.

Related: OSPRH-13415

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>