Add CA cert to bootstrap service
We want to copy the combined-ca-cert-bundle to the compute nodes to
ensure that any third party certs are trusted.  Lets use the bootstrap
service to do this.  This change is paired with - but does not depend on
a change in openstack-operator that will ensure that the bootstrap service has the
cert bundle mounted to /var/lib/openstack/cacerts/bootstrap.

The relevant change is https://github.com/openstack-k8s-operators/openstack-operator/pull/1374

Jira: OSPRH-14205