Consolidate "redhat" service conditions
A node can be running RHEL, but doesn't necessarly need to
register/subscribe.

This PR ensures the node is running RHEL (as before), but also checks we
have at least `rhc_auth` parameter defined, and not empty.
This should provide enough security to re-add the "redhat" service in
the default list.

Related-Issue: OSPRH-15644