Add a validation against edpm_network_config_template
Depending on the tool used to configure network, the configuration
template is different.

This patch adds a quick validation to avoid obvious failures, by
ensuring a specific key is present in the template.

If the admin is considering "nmstate", "interfaces" key must be
present.

If the admin is considering "ifcfg", "network_config" key must be
present.