Add variable to override all download retries
Currently each edpm_* role defines their own set of variables
to control the number of retries and delay between retries
when attempting to download and install the system packages.

There is over 20 such roles that involve it already, however
there is no way to set those variables easily for all roles
at once – the only option for someone is to set all about
40 variables in worst scenario.

This commit introduces two variables that would be used
as first choice, if provided, by all the roles to allow
the setting the download retries much easier.