Explicitly install sos report in EDPM nodes
It might happen that the sos package is not installed by default in a
system where it doesn't appear as a dependency for other packages.
For this reason this patch makes it part of the default EDPM bootstrap
packages to explicitly install it if not present.

Jira: https://issues.redhat.com/browse/OSPRH-17414

Signed-off-by: Francesco Pantano <fpantano@redhat.com>