Gather local ansible facts if not done for roles
While the gathering of ansible local facts has been added to all
playbooks that use roles where the facts are required, this commit adds
it to the individual roles as well.

This is needed for the molecule tests that use the roles directly and
not the playbooks, and likewise to enable the roles to operate when used
standalone outside of the playbooks.

Signed-off-by: James Slagle <jslagle@redhat.com>