Move some vars from internal role vars to default
vars/main.yaml is for internal role variables and have a higher
prededence. They can't be overriden by inventory vars. Let's move
them to default/main.yaml.

We don't see this issue in 17.1 because the tripleo_sshd role has
a vars/redhat.yaml which is used instead of vars/main.yaml and it
does not have those role vars.

Signed-off-by: rabi <ramishra@redhat.com>