Configure LVM's devices file
Configure LVM's "devices file" in order to prevent it from scanning
cinder volume attachments.

LVM uses the concept of a devices file to control whether it
scans local and dynamically attached (SAN) volumes. This is an
enhancement over using LVM device filters. The devices file
essentially provides a whitelist of devices that should be scanned,
and the file is automatically updated by LVM commands such as
'lvcreate'. For this to work, the devices file must exist even if
it's empty.

Resolves: OSPRH-20217
Signed-off-by: Alan Bishop <abishop@redhat.com>