Make config directories consistent across services
Some services used edpm_container_standalone and others directly use
edpm_container_manage and the way we generate the config files and use
them is spread across multiple locations which is confusing and difficult
to troubleshoot. This sanitizes the locations.

This is required to identify all the services deployed vs to be deployed
and remove orphaned containers/services.

Note: This does not make the changes for edpm_nova role which would be
handled in a separate PR.

jira: https://issues.redhat.com/browse/OSPRH-19243

Signed-off-by: rabi <ramishra@redhat.com>