Add support for the sidecar containers in ``neutron_ovn`` role
Similarly to what was done in Tripleo, neutron agents which are spawning
some long running processes, like haproxy should run them in the sidecar
containers to avoid breakege in the dataplane when e.g. agent's container
is restarted.

This patch adds wrapper for haproxy sidecar in the neutron_metadata role.
This wrapper script is taken almost directly from the TripleO's
puppet module [1] and is just converted to jinja2 format and to support
only Podman as container_cli.

This patch adds also support for kill-scripts which is used to stop
and clean sidecar containers. It is takend almost directly from the
TripleO Heat Templates [2].

This patch is an adaptation of [3].

[1] https://opendev.org/openstack/puppet-tripleo/src/branch/master/templates/neutron/haproxy.epp
[2] https://github.com/openstack/tripleo-heat-templates/blob/master/deployment/neutron/kill-script

Resolves: OSPRH-22667
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I3824863338447a667cbf175265e3ad43ad079d02