edpm-podman: clean up old images in edpm_container_manage
This change modifies the podman image cleanup process to avoid removing
images that are actively used. Instead of using 'podman
image prune -a -f' which could cause service outages, we now clean up
images in edpm_container_manage and purge task is only kept for manual
cleanup.

The until (144h = 6d) value is based on expectation of maintenance
happening every week so we shoud remove containers and images from the
previous maintenance. It prevents the concurrency issues as we don't
need to worry to remove containers under us while starting them so
minimal value should be 20m.

Note: the label name is imperfect and we should consider collecting
more labels if name is used or we risk picking up on the new container
too. This is another reason to use higher until value.

Related: OSPRH-22122 OSPRH-23128
Closes: OSPRH-17111
Signed-off-by: Lukas Bezdicka <lbezdick@redhat.com>