Restart edpm_multipathd.service
Stopping multpathd.service and socket is fixed by cbc084a.
However, stopping host side multipathd.service and socket broke the
communication with multipath and multipathd commands.
As a result, nova cannot start an instance.

To fix the communication, restarting edpm_mulitpathd.service is
required if it's running.

This fix is needed for FR4 branch only because FR5 removes containerized
iscsid and mulitpathd.

Closes: OSPRH-23113

Signed-off-by: Keigo Noha <knoha@redhat.com>